requirements.txt
imgbeddings
mediapipe
opencv-contrib-python
requests
flask
flask-cors
pyqt5
numpy
pillow